requirements.txt
winwifi #https://github.com/mrjohannchang/winwifi?tab=readme-ov-file
pillow